requirements.txt
paho-mqtt
pika
pymongo
python-dotenv